Update logging to follow best practice
One should not f-string logging - it has built-in handling for string
formatting, which only runs if the logging is actually going to happen.
This prevents formatting f-strings for log messages which are suppressed
due to logging levels.  Also, the parameters will be available for use in
non-text logging such as a JSON handler.

Also switch to using `logging.exception()` when in an exception handler,
which will include the traceback data.